Move PerformanceTestSuite to performance tests bundle
It's only used there and as only this one is run by perf tests it's of
no use in general tests (and really not used there).
