+++ org.eclipse.jdt.core.tests.performance/src/org/eclipse/jdt/core/tests/performance/AllPerformanceTests.java
@@ -16,7 +16,6 @@ package org.eclipse.jdt.core.tests.performance;
 import java.lang.reflect.*;
 import java.text.NumberFormat;
 
-import org.eclipse.jdt.core.tests.junit.extension.PerformanceTestSuite;
 import org.eclipse.jdt.core.tests.junit.extension.TestCase;
 import junit.framework.Test;
